requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.8
fpdf==1.7.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langsmith==0.3.10
lxml==5.3.1
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.64.0
orjson==3.10.15
packaging==24.2
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
PyMuPDF==1.25.3
pypdf==5.3.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tenacity==9.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
zstandard==0.23.0
